Update primitive_calculator
Given an integer n, compute the minimum number of operations needed to obtain the number n starting from the number 1. 

The input consists of a single integer 1 ≤ n ≤ 106. Output Format. In the ﬁrst line, output the minimum number k of operations needed to get n from 1. In the second line output a sequence of intermediate numbers. That is, the second line should contain positive integers a0,a2,...,ak−1 such that a0 = 1, ak−1 = n and for all 0 ≤ i < k−1, ai+1 is equal to either ai + 1, 2ai, or 3ai. If there are many such sequences, output any one of them